Fixed retina and mask obs space to be uint8 and int32

+++ real_robots/envs/robot.py
@@ -87,13 +87,13 @@ class Kuka(URDFBasedRobot):
                 self.ObsSpaces.TOUCH_SENSORS: gym.spaces.Box(
                     0, np.inf, [self.num_touch_sensors], dtype=float),
                 self.ObsSpaces.RETINA: gym.spaces.Box(
-                    0, 255, [Kuka.eye_height, Kuka.eye_width, 3], dtype=float),
+                    0, 255, [Kuka.eye_height, Kuka.eye_width, 3], dtype=np.uint8),
                 self.ObsSpaces.GOAL: gym.spaces.Box(
-                    0, 255, [Kuka.eye_height, Kuka.eye_width, 3], dtype=float),
+                    0, 255, [Kuka.eye_height, Kuka.eye_width, 3], dtype=np.uint8),
                 self.ObsSpaces.MASK: gym.spaces.Box(
-                    0, 255, [Kuka.eye_height, Kuka.eye_width], dtype=float),
+                    0, 255, [Kuka.eye_height, Kuka.eye_width], dtype=np.int32),
                 self.ObsSpaces.GOAL_MASK: gym.spaces.Box(
-                    0, 255, [Kuka.eye_height, Kuka.eye_width], dtype=float),
+                    0, 255, [Kuka.eye_height, Kuka.eye_width], dtype=np.int32),
                 self.ObsSpaces.OBJ_POS: gym.spaces.Dict(obj_obs)
                 }
             )
